fix: remove sensible data in versionated code
