+++ .env.example
@@ -1,0 +1,4 @@
+VTOOLS_API="https://events.vtools.ieee.org/RST/events/api/public/<vtools_event_api_version>/"
+CALENDAR_ID="<google_calendar_collab_email>"
+COUNTRY_ID=<vtools_events_country_id>
+FILTER_MODE="COUNTRY"
