+++ .gitignore
@@ -1,4 +1,6 @@
 .venv/
+.env
 credentials.json
 token.json
 __pycache__/
+.idea
