+++ main.py
@@ -1,19 +1,23 @@
+import os
 import lib.vtools_fetcher as vtf
 from lib.calendar_api import CalendarAPI
+from dotenv import load_dotenv
 
-COLOMBIA_SECTION_ID = 45
+load_dotenv()
+
+country_id = os.getenv('COUNTRY_ID')
+
+print("Fetching events from vTools")
+events = vtf.get_events_by_country(country_id)
+print(f"{len(events)} Events fetched")
 
 print("Bootstraping CalendarAPI")
 calendar = CalendarAPI()
 print("Bootstraped CalendarAPI")
 
-print("Fetching events from vTools")
-colombia_events = vtf.get_events_by_country(COLOMBIA_SECTION_ID)
-print(f"{len(colombia_events)} Events fetched")
-
 stats = {'added': 0, 'updated': 0, 'no_change': 0}
 print("Adding/Updating events to Google Calendar")
-for event in colombia_events:
+for event in events:
     result = calendar.add_or_update_event(event)
     stats[result] += 1
